[update] Refactor, Re-Order functions
